requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
pillow==10.2.0
openai==1.12.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.27.0
psycopg2-binary==2.9.9
